@@ -9,11 +9,11 @@
 # from scratch. The latter is a flawed and unsustainable approach (the more migrations
 # you'll amass, the slower it'll run and the greater likelihood for issues).
 #
-# It's strongly recommended to check this file into your version control system.
+# It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20130107010733) do
+ActiveRecord::Schema.define(version: 20130107010733) do
 
-  create_table "orders", :id => false, :force => true do |t|
+  create_table "orders", id: false, force: true do |t|
     t.string   "token"
     t.string   "transaction_id"
     t.string   "address_one"
@@ -32,26 +32,26 @@ ActiveRecord::Schema.define(:version => 20130107010733) do
     t.string   "phone"
     t.string   "name"
     t.date     "expiration"
-    t.datetime "created_at",        :null => false
-    t.datetime "updated_at",        :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
     t.integer  "payment_option_id"
   end
 
-  create_table "payment_options", :force => true do |t|
+  create_table "payment_options", force: true do |t|
     t.decimal  "amount"
     t.string   "amount_display"
     t.text     "description"
     t.string   "shipping_desc"
     t.string   "delivery_desc"
     t.integer  "limit"
-    t.datetime "created_at",     :null => false
-    t.datetime "updated_at",     :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
   end
 
-  create_table "users", :force => true do |t|
+  create_table "users", force: true do |t|
     t.string   "email"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
   end
 
 end
