@@ -13,7 +13,7 @@
 //= require html5shiv-printshiv
 //= require jquery
 //= require jquery_ujs
-//= require jquery-ui
+//= require jquery.ui.all
 //= require jquery.details
 //= require jquery.textchange
 //= require preorder
